# Router: PoC stack based backtracking