PoC regarding issue #1754