@stffabi 2021-02-10
I've further improved the PoC and was able to remove almost all corner cases in the algorithm, which have been introduced to fix routing issues.