@lammel 2021-02-11
@stffabi Thanks, very appreciated,

There is quite some performance penality involved which should be reviewed:

```
RouterStaticRoutes-2                      21.7µs ± 1%    23.1µs ± 2%   +6.09%  (p=0.000 n=7+8)
RouterStaticRoutesMisses-2                 705ns ± 4%    1090ns ± 6%  +54.65%  (p=0.000 n=8+8)
RouterGitHubAPI-2                         41.4µs ± 4%    43.6µs ± 6%   +5.25%  (p=0.007 n=8+8)
RouterGitHubAPIMisses-2                    860ns ± 4%    1186ns ± 4%  +37.91%  (p=0.000 n=8+8)
RouterParseAPI-2                          2.17µs ± 3%    2.89µs ± 3%  +33.34%  (p=0.000 n=7+8)
RouterParseAPIMisses-2                     488ns ± 5%     877ns ± 5%  +79.57%  (p=0.000 n=8+8)
RouterGooglePlusAPI-2                     1.41µs ± 2%    1.65µs ± 2%  +16.76%  (p=0.000 n=8+7)
RouterGooglePlusAPIMisses-2                725ns ± 3%    1037ns ± 3%  +43.03%  (p=0.000 n=8+7)
RouterParamsAndAnyAPI-2                   3.37µs ± 5%    3.79µs ± 4%  +12.58%  (p=0.000 n=8+8)
```

The above is from the automatic benchstat, so might be little off. But the numbers do indicate a performance impact that should be reviewed. 
Maybe you can have a look at that.