@stffabi 2021-02-11
@lammel You're very welcome.

Yes I know there are some performance penalties. But as already mentioned it is in a PoC state and my first priority was to have a working and correct algorithm before introducing any optimizations.
We have to make sure which test cases really could and should get on par with master. There might be cases were we could never reach the current performance anymore, because the current algorithm on master too early stops and responds with not found and doesn't find all routes.

One question regarding the base of the benchstat of a commit. What's the base of the automatic benchstat of a commit? Is the `old` version always the latest commit on the master branch from with the PR has been diverged? Is every commit on the PR compared with the divergence commit on master, or is the base always the previous commit?

I would be glad to invest some more time to improve the performance, if the echo community plans to have this merged. Unfortunately I had some bad experiences with other open source projects (not with echo 😄 ) lately, where I ended up with invested time and PRs that never got merged.