@stffabi 2021-02-12
I've improved the PoC, it now doesn't anymore store the backtrack information on the heap but on the goroutine's stack. As a consequence the maximum level of nested backtracking needs to be known beforehand during compile-time. Currently as this is PoC, the `Find` simply panics if the level is too deep, but we could easily check this during the registration of the routes. Therefore it would panic beforehand and not depending on the incoming request. But let's first focus on the performance.

I've started to experiment around with different values for the max backtracking depth. First started with 1 (which would be identical with the current implementation) and then I've successively increased it. Please note in order that all benchmark tests run and don't panic we need at least a max depth of 2. 

Here are the results (benchmarks done on my local machine), delta is always compared to old (932976ded6e22eb0d80f73c540aef0849b2047af)

```
name                           old time/op    new-1 time/op    delta    new-2 time/op    delta    new-3 time/op    delta    new-4 time/op    delta    new-5 time/op    delta    new-8 time/op    delta    new-12 time/op    delta
RouterStaticRoutes-8           10.6µs ± 5%       8.8µs ± 3%  -16.88%       8.7µs ± 1%  -17.24%       8.8µs ± 3%  -16.82%       8.8µs ± 1%  -16.85%       9.1µs ± 3%  -13.57%       9.1µs ± 2%  -13.49%        9.6µs ± 6%   -8.86%
RouterStaticRoutesMisses-8      394ns ± 2%       393ns ± 1%     ~          402ns ± 1%   +2.13%       405ns ± 2%   +2.93%       409ns ± 2%   +3.90%       421ns ± 1%   +6.82%       440ns ± 7%  +11.85%        461ns ± 2%  +17.02%
RouterGitHubAPI-8              23.5µs ± 6%                                19.4µs ± 1%  -17.63%      19.3µs ± 2%  -17.70%      19.5µs ± 1%  -17.17%      19.9µs ± 2%  -15.30%      20.0µs ± 2%  -14.70%       20.6µs ± 2%  -12.14%
RouterGitHubAPIMisses-8         502ns ± 6%                                 460ns ± 1%   -8.37%       464ns ± 3%   -7.43%       466ns ± 2%   -7.16%       487ns ± 6%   -3.01%       490ns ± 1%     ~           513ns ± 2%   +2.25%
RouterParseAPI-8               1.28µs ± 1%                                1.14µs ± 2%  -11.00%      1.14µs ± 1%  -11.10%      1.14µs ± 1%  -10.53%      1.17µs ± 2%   -8.90%      1.20µs ± 2%   -6.48%       1.23µs ± 1%   -3.68%
RouterParseAPIMisses-8          280ns ± 5%                                 278ns ± 3%     ~          281ns ± 4%     ~          281ns ± 2%     ~          291ns ± 2%   +3.70%       309ns ± 3%  +10.16%        329ns ± 3%  +17.17%
RouterGooglePlusAPI-8           827ns ± 5%                                 715ns ± 1%  -13.57%       722ns ± 2%  -12.69%       723ns ± 2%  -12.62%       735ns ± 0%  -11.11%       751ns ± 2%   -9.15%        769ns ± 2%   -7.02%
RouterGooglePlusAPIMisses-8     414ns ± 4%                                 368ns ± 2%  -11.29%       370ns ± 3%  -10.84%       373ns ± 2%  -10.05%       385ns ± 2%   -7.02%       402ns ± 3%   -2.95%        422ns ± 1%   +1.85%
RouterParamsAndAnyAPI-8        1.84µs ± 2%                                1.53µs ± 1%  -17.05%      1.54µs ± 1%  -16.39%      1.53µs ± 1%  -16.67%      1.57µs ± 1%  -14.97%      1.58µs ± 1%  -14.13%       1.62µs ± 2%  -12.13%
```

As one can see, the new algorithm with a max depth of 2 outperforms the current version in almost any benchmark. Furthermore one can see that by increasing the max depth the performance begins to degrade, especially for the `Misses` benchmarks. As increasing the max depth doesn't have any influence on the amount of processing our algorithm has to do (we only need a depth of 2 for successfully run the benchmarks), the performance penalty comes from the go runtime in order to allocate the goroutine's stack. So for increasing the max depth the go runtime needs to allocate more memory on the goroutine's stack, which is obviously clearly out of our control.

So what could we conclude from this: One has to balance between the max depth of backtracking echo could support and the performance penalty one is willing to pay for it.

I currently see no options to further improve this. First we have to store the backtracking information somewhere, 
which is now already stored on the fastest memory section, the goroutine's stack. Second we need memory in order to store multiple levels, so memory needs to be allocated and we can't get any better than letting the go runtime allocating this memory for us on the goroutine's stack.

The only options I could see, but those options could make the code harder to maintain, are the following:
- Having different implementations of `Find` with different max depths.
- Chose the most suitable `Find` for the current max backtracking depth, as determined during route registration.
- Maybe having a fallback to switch to a heap implementation if the current max backtracking depth exceeds a threshold.

This would allow echo to work with the best performance depending on the nested backtracking depth, that the user has registered on the router.