@aldas 2021-02-12
I think @stffabi version is quite good fix. Performance hit on backtracking can not be avoided as by algoritm backtracking is done until suitable node is found up in tree to the right or we are back at root node and there are nowhere else to go.

@stffabi version is quite nice as it simplifies current implementation quite a bit and makes it more approachable.  If that depth calculation is moved to `route.Insert()` and panic is remove (`not found` result would be fine by me - better than current) and if possible stack and pop/push moved to separate type this change looks quite ok.

I tried to implement same algorithm but without stack (as I thought it would definitely mean allocating) and result are similar - misses (backtracking) is always performance hit.



My solution https://github.com/aldas/echo/blob/issue-1754_router_backtracking/router.go#L552

And comparison to 

* old.out - current echo `Find()` implementation
* stffabi.out - @stffabi stack based implementation
* new.out - mine (using direction for tree traversial)

```bash
x@x:~/code/echo$ benchstat old.out stffabi.out 
name                         old time/op    new time/op    delta
RouterStaticRoutes-6           16.0µs ± 1%    15.1µs ± 3%   -5.12%  (p=0.000 n=20+19)
RouterStaticRoutesMisses-6      515ns ± 1%     615ns ± 3%  +19.48%  (p=0.000 n=20+19)
RouterGitHubAPI-6              30.1µs ± 3%    33.0µs ±15%   +9.68%  (p=0.000 n=20+20)
RouterGitHubAPIMisses-6         621ns ± 0%     678ns ± 5%   +9.15%  (p=0.000 n=16+20)
RouterParseAPI-6               1.61µs ± 2%    1.55µs ± 3%   -3.64%  (p=0.000 n=20+18)
RouterParseAPIMisses-6          349ns ± 1%     404ns ± 1%  +15.81%  (p=0.000 n=20+17)
RouterGooglePlusAPI-6          1.04µs ± 0%    1.01µs ± 3%   -3.36%  (p=0.000 n=15+20)
RouterGooglePlusAPIMisses-6     516ns ± 1%     547ns ± 4%   +6.00%  (p=0.000 n=19+20)
RouterParamsAndAnyAPI-6        2.44µs ± 1%    2.34µs ± 6%   -4.31%  (p=0.000 n=20+20)

x@x:~/code/echo$ benchstat new.out stfabi.out 
name                         old time/op    new time/op    delta
RouterStaticRoutes-6           14.9µs ± 1%    15.1µs ± 3%   +1.77%  (p=0.000 n=18+19)
RouterStaticRoutesMisses-6      585ns ± 2%     615ns ± 3%   +5.20%  (p=0.000 n=20+19)
RouterGitHubAPI-6              31.2µs ± 1%    33.0µs ±15%   +5.69%  (p=0.033 n=20+20)
RouterGitHubAPIMisses-6         763ns ± 1%     678ns ± 5%  -11.15%  (p=0.000 n=20+20)
RouterParseAPI-6               1.38µs ± 3%    1.55µs ± 3%  +12.23%  (p=0.000 n=20+18)
RouterParseAPIMisses-6          394ns ± 1%     404ns ± 1%   +2.58%  (p=0.000 n=16+17)
RouterGooglePlusAPI-6           887ns ± 2%    1009ns ± 3%  +13.69%  (p=0.000 n=19+20)
RouterGooglePlusAPIMisses-6     616ns ± 0%     547ns ± 4%  -11.08%  (p=0.000 n=14+20)
RouterParamsAndAnyAPI-6        2.43µs ± 2%    2.34µs ± 6%   -3.89%  (p=0.000 n=20+20)

x@x:~/code/echo$ benchstat old.out new.out 
name                         old time/op    new time/op    delta
RouterStaticRoutes-6           16.0µs ± 1%    14.9µs ± 1%   -6.77%  (p=0.000 n=20+18)
RouterStaticRoutesMisses-6      515ns ± 1%     585ns ± 2%  +13.58%  (p=0.000 n=20+20)
RouterGitHubAPI-6              30.1µs ± 3%    31.2µs ± 1%   +3.78%  (p=0.000 n=20+20)
RouterGitHubAPIMisses-6         621ns ± 0%     763ns ± 1%  +22.86%  (p=0.000 n=16+20)
RouterParseAPI-6               1.61µs ± 2%    1.38µs ± 3%  -14.14%  (p=0.000 n=20+20)
RouterParseAPIMisses-6          349ns ± 1%     394ns ± 1%  +12.90%  (p=0.000 n=20+16)
RouterGooglePlusAPI-6          1.04µs ± 0%    0.89µs ± 2%  -14.99%  (p=0.000 n=15+19)
RouterGooglePlusAPIMisses-6     516ns ± 1%     616ns ± 0%  +19.20%  (p=0.000 n=19+14)
RouterParamsAndAnyAPI-6        2.44µs ± 1%    2.43µs ± 2%   -0.43%  (p=0.014 n=20+20)
```


also I think all implementation have atm 1 failing testcase. When backtracking from `/:name/lastname` dead end `/:name` should be the match has it has priority over `/*`
```go
func TestRouterBacktrackingFromParamAny2(t *testing.T) {
	e := New()
	r := e.router

	r.Add(http.MethodGet, "/*", handlerHelper("case", 1))
	r.Add(http.MethodGet, "/:name", handlerHelper("case", 2))
	r.Add(http.MethodGet, "/:name/lastname", handlerHelper("case", 3))

	c := e.NewContext(nil, nil).(*context)
	
	r.Find(http.MethodGet, "/firstname/", c) // trailing slash confuses param kind matching algorithm
	c.handler(c)
	assert.Equal(t, "/:name", c.Get("path"))  // FIXME: currently matches `/*` which is low priority than `/:name` and therefore should not be matched
	assert.Equal(t, "firstname/", c.Param("*"))


	r.Find(http.MethodGet, "/firstname/test", c)
	c.handler(c)
	assert.Equal(t, "/:name", c.Get("path")) // FIXME:  currently matches `/*`
	assert.Equal(t, "firstname/test", c.Param("*"))
}